+++ src/pages/blog/[id].tsx
@@ -5,7 +5,7 @@ import Link from 'next/link';
 import styled from 'styled-components';
 import { toJapanDate } from '@/libs/dayjs';
 import { microcms } from '@/libs/microcms';
-import { Post, Blog } from '@/utils/types/blog';
+import { Post, Blog } from '@/types/blog';
 import MaxWidthWrapper from 'src/components/ui/MaxWidthWrapper';
 
 const OGP_URL = 'https://og-image-takuyakikuchi.vercel.app';
