+++ src/pages/blog/page/[page].tsx
@@ -1,8 +1,8 @@
 import { GetStaticProps, GetStaticPaths } from 'next';
 import { useRouter } from 'next/router';
 import { microcms } from '@/libs/microcms';
+import { Blog, Post } from '@/types/blog';
 import { PER_PAGE } from '@/utils/constants';
-import { Blog, Post } from '@/utils/types/blog';
 import Home from 'src/pages/index';
 
 const parsePageParams = (page: string | string[] | undefined) => {
