+++ src/pages/index.tsx
@@ -2,7 +2,7 @@ import { GetStaticProps } from 'next';
 import Head from 'next/head';
 import styled from 'styled-components';
 import { microcms } from '@/libs/microcms';
-import { Blog, Post } from '@/utils/types/blog';
+import { Blog, Post } from '@/types/blog';
 import Card from 'src/components/home/Card';
 import { Pagination } from 'src/components/home/Pagination';
 import MaxWidthWrapper from 'src/components/ui/MaxWidthWrapper';
